refactoring our test suite
