+++ tests/api/subscriptions.rs
@@ -1,0 +1,45 @@
+use crate::helpers::*;
+
+#[tokio::test]
+async fn subscribe_returns_200() {
+    let app: TestApp = spawn_app().await;
+
+    let body = "name=le%20guin&email=ursula_le_guin%40gmail.com";
+    let response = app.post_subscriptions(body.into()).await;
+    
+    assert_eq!(200, response.status().as_u16());
+
+    //db query
+    let saved = sqlx::query!("SELECT email, name FROM subscriptions")
+        .fetch_one(&app.db_pool)
+        .await
+        .expect("Failed to fetch saved subscription");
+
+    println!("{}", saved.email);
+    println!("{}", saved.name);
+
+    assert_eq!(saved.email, "ursula_le_guin@gmail.com");
+    assert_eq!(saved.name, "le guin");
+}
+
+#[tokio::test]
+async fn subscribe_returns_400() {
+    let app = spawn_app().await;
+
+    let test_cases = vec![
+        ("name=&email=ursula_le_guin%40gmail.com", "empty name"),
+        ("name=Ursula&email=", "empty email"),
+        ("name=Ursula&email=definitely-not-an-email", "invalid email")
+    ];
+
+    for (body, error_message) in test_cases {
+        let response = app.post_subscriptions(body.into()).await;
+
+        assert_eq!(
+            400,
+            response.status().as_u16(),
+            "The API did not fail with 400 Bad Request when the payload was {}.",
+            error_message
+        );
+    }
+}
