--- tests/health_check.rs
@@ -1,179 +1,0 @@
-use std::net::TcpListener;
-use sqlx::{
-    PgPool,
-    PgConnection,
-    Connection,
-    Executor
-};
-use uuid::Uuid;
-use email_newsletter::{
-    configuration::DatabaseSettings,
-    telemetry::*,
-    email_client::EmailClient
-};
-use once_cell::sync::Lazy;
-
-static TRAICING: Lazy<()> = Lazy::new(|| {
-    let default_filter_level = "info".to_string();
-    let subscriber_name = "test".to_string();
-
-    if std::env::var("TEST_LOG").is_ok() {
-        let subscriber = get_subscriber(
-            subscriber_name,
-            default_filter_level,
-            std::io::stdout
-        );
-        init_subscriber(subscriber);
-    } else {
-        let subscriber = get_subscriber(
-            subscriber_name,
-            default_filter_level,
-            std::io::sink
-        );
-        init_subscriber(subscriber);
-    }
-
-});
-
-struct TestApp {
-    pub address: String,
-    pub db_pool: PgPool
-}
-
-#[tokio::test]
-async fn health_check_works() {
-    // Arrange
-    let app = spawn_app().await;
-    let client = reqwest::Client::new();
-    //We use reqwest to perform HTTP request against our application
-    
-    println!("{}", &app.address);
-    
-    //Act
-    let response = client
-        .get(format!("{}/health_check", &app.address))
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    //Assert
-    assert!(response.status().is_success());
-}
-
-#[tokio::test]
-async fn subscribe_returns_200() {
-    let test_app: TestApp = spawn_app().await;
-
-    let client = reqwest::Client::new();
-
-    let body = "name=le%20guin&email=ursula_le_guin%40gmail.com";
-    let response = client
-        .post(&format!("{}/subscriptions", &test_app.address))
-        .header("Content-Type", "application/x-www-form-urlencoded")
-        .body(body)
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    assert_eq!(200, response.status().as_u16());
-
-    //db query
-    let saved = sqlx::query!("SELECT email, name FROM subscriptions")
-        .fetch_one(&test_app.db_pool)
-        .await
-        .expect("Failed to fetch saved subscription");
-
-    println!("{}", saved.email);
-    println!("{}", saved.name);
-
-    assert_eq!(saved.email, "ursula_le_guin@gmail.com");
-    assert_eq!(saved.name, "le guin");
-}
-
-#[tokio::test]
-async fn subscribe_returns_400() {
-    let app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    let test_cases = vec![
-        ("name=&email=ursula_le_guin%40gmail.com", "empty name"),
-        ("name=Ursula&email=", "empty email"),
-        ("name=Ursula&email=definitely-not-an-email", "invalid email")
-    ];
-
-    for (body, error_message) in test_cases {
-        let response = client
-            .post(&format!("{}/subscriptions", &app.address))
-            .header("Content-Type", "application/x-www-form-urlencoded")
-            .body(body)
-            .send()
-            .await
-            .expect("Failed to execute request");
-
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            "The API did not fail with 400 Bad Request when the payload was {}.",
-            error_message
-        );
-    }
-}
-
-async fn spawn_app() -> TestApp {
-    Lazy::force(&TRAICING);
-
-    let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
-    let port = listener.local_addr().unwrap().port();
-    let address = format!("http://127.0.0.1:{}", port);
-
-    let mut configuration = email_newsletter::configuration::get_configuration().expect("Failed to get the configuration file");
-    configuration.database.database_name = Uuid::new_v4().to_string();
-    let db_pool = configure_database(&configuration.database).await;
-    let timeout = configuration.email_client.timeout();
-
-    let sender_email = configuration.email_client.sender().expect("Invalid sender email address");
-    let email_client = EmailClient::new(
-        configuration.email_client.base_url.clone(),
-        sender_email,
-        configuration.email_client.authorization_token,
-        timeout
-    );
-
-    let server = email_newsletter::startup::run(listener, db_pool.clone(), email_client).expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-
-    //We return the application address to the caller
-    TestApp {
-        address,
-        db_pool
-    }
-}
-
-pub async fn configure_database(config: &DatabaseSettings) -> PgPool {
-    let mut connection = PgConnection::connect_with(&config.without_db())
-    .await
-    .expect("Failed to connect to Postgres");
-
-    connection
-        .execute(format!(r#"CREATE DATABASE "{}";"#, &config.database_name).as_str())
-        .await
-        .expect("Failed to create database");
-
-    //Migrate database
-    let connection_pool = PgPool::connect_with(config.with_db())
-        .await
-        .expect("Failed to connect to Postgres.");
-
-    sqlx::migrate!("./migrations")
-        .run(&connection_pool)
-        .await
-        .expect("Failed to migrate the database");
-
-    connection_pool
-}
-
-#[test]
-fn dummy_fail() {
-    let result: Result<&str, &str> = Err("The app crashed due to an IO error");
-    claims::assert_ok!(result);
-}
